Revert "接口实现"
This reverts commit 017a00af5ac5c67bd46a43ee5cfcf663c57cfbea.

+++ src/java/com/yizhiyun/dingcan/service/VendorService.java
@@ -7,6 +7,7 @@
 package com.yizhiyun.dingcan.service;
 
 import com.yizhiyun.dingcan.dao.BaseDao;
+import com.yizhiyun.dingcan.model.Customer;
 import com.yizhiyun.dingcan.model.Vendor;
 import java.util.HashMap;
 import java.util.List;
@@ -32,7 +33,7 @@ public class VendorService {
         return baseDao.getList("VendorMapper.getVendorsByTel", telNo);
     }
 
-    public List<Vendor> getVendorsByCoordinate(double minLatitude, double maxLatitude, double minLongitude, double maxLongitude) {
+    public List<Vendor> getVendorsByCoordinate(float minLatitude, float maxLatitude, float minLongitude, float maxLongitude) {
         Map<String, Object> params = new HashMap<String, Object>();
         params.put("minLatitude", minLatitude);
         params.put("minLongitude", minLongitude);
